# db: remove levelIterBoundaryContext and rangeDelIter pointer shenanigans