#### db: remove levelIterBoundaryContext pointer shenanigans

The level iterator can optionally populate an *external*
levelIterBoundaryContext through a pointer; this is set to point to
the corresponding field in the merging iterators.

This subtlety of how this field is updated is unnecessary. We can make
the level iterator always keep track of this information internally;
in the merging iterator we initialize a pointer to this state in the
level iterator or to an empty struct if we are not using a level
iterator.

#### db: rework the mechanism to support a changing rangeDelIter

When using a `levelIter`, the `rangeDelIter` changes with each file.
The current mechanism for higher levels (in particular the merging
iterator) to obtain the current `rangeDelIter` is to configure the
`levelIter` with a pointer to a field. This is very subtle, as the
field seems to "magically" change from the perspective of the higher
level code.

This commit changes this mechanism: we define a `rangeDelIterHolder`
which can be configured either with a fixed `rangeDeliter` or with a
`levelIter` that is configured to expose the current `rangeDelIter`. A
`get()` method returns either the fixed `rangeDelIter` or the
`levelIter`'s current `rangeDelIter`.