@RaduBerinde 2023-03-03
I added a commit that performs a similar treatment to the `rangeDelIter`. PTAL. Benchmarks show no significant differences (this is master vs this PR):

```
name                                                                                                          old time/op  new time/op  delta
NewIterReadAmp/10                                                                                              708ns ± 3%   727ns ± 1%     ~     (p=0.114 n=4+4)
NewIterReadAmp/100                                                                                            19.6µs ± 0%  21.6µs ± 1%  +10.28%  (p=0.029 n=4+4)
NewIterReadAmp/1000                                                                                            222µs ± 1%   229µs ± 0%   +3.05%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=1/forward-only=false                                  24.4µs ± 2%  24.8µs ± 3%     ~     (p=0.343 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=1/forward-only=true                                   14.7µs ± 1%  14.7µs ± 1%     ~     (p=0.486 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=10/forward-only=false                                  313µs ± 2%   308µs ± 1%     ~     (p=0.114 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=10/forward-only=true                                   100µs ± 1%   103µs ± 0%   +2.46%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=100/forward-only=false                                2.21ms ± 1%  2.21ms ± 1%     ~     (p=0.686 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=100/forward-only=true                                  890µs ± 0%   915µs ± 1%   +2.78%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=1/forward-only=false                                1.68ms ± 0%  1.71ms ± 1%   +2.08%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=1/forward-only=true                                  510µs ± 1%   490µs ± 1%   -3.93%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=10/forward-only=false                               37.6ms ± 0%  37.4ms ± 0%   -0.67%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=10/forward-only=true                                1.50ms ± 2%  1.56ms ± 1%   +4.16%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=100/forward-only=false                               330ms ± 0%   338ms ± 1%   +2.48%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=100/forward-only=true                               9.53ms ± 2%  9.87ms ± 2%     ~     (p=0.057 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=1/forward-only=false                                183ms ± 1%   185ms ± 1%     ~     (p=0.057 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=1/forward-only=true                                5.36ms ± 2%  5.10ms ± 0%   -4.83%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=10/forward-only=false                               341ms ± 3%   327ms ± 0%   -4.13%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=10/forward-only=true                               13.6ms ± 1%  14.1ms ± 0%   +3.33%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=100/forward-only=false                              4.03s ± 1%   3.88s ± 1%   -3.54%  (p=0.029 n=4+4)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=100/forward-only=true                              80.1ms ± 1%  80.6ms ± 2%     ~     (p=0.686 n=4+4)
IteratorSeekGE                                                                                                 428ns ± 4%   414ns ± 6%     ~     (p=0.200 n=4+4)
IteratorNext                                                                                                  16.5ns ± 1%  16.2ns ± 1%   -1.83%  (p=0.029 n=4+4)
IteratorPrev                                                                                                  21.9ns ± 1%  22.4ns ± 1%   +2.35%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=false                        247ns ± 1%   245ns ± 0%   -1.11%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=false/with-tombstone=true                         214ns ± 0%   219ns ± 0%   +2.73%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=false                         441ns ± 1%   436ns ± 0%     ~     (p=0.057 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=false/bloom=true/with-tombstone=true                          255ns ± 0%   252ns ± 0%   -1.18%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=false                         287ns ± 0%   282ns ± 0%   -1.78%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=false/with-tombstone=true                          223ns ± 0%   232ns ± 0%   +4.03%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=false                          449ns ± 1%   441ns ± 2%   -1.75%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=1/two-level=true/bloom=true/with-tombstone=true                           261ns ± 0%   257ns ± 1%   -1.34%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=false                        249ns ± 1%   250ns ± 0%     ~     (p=0.343 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=false/with-tombstone=true                         216ns ± 0%   223ns ± 0%   +3.34%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=false                         444ns ± 1%   454ns ± 5%     ~     (p=0.114 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=false/bloom=true/with-tombstone=true                          256ns ± 0%   257ns ± 0%     ~     (p=0.057 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=false                         287ns ± 0%   288ns ± 0%   +0.52%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=false/with-tombstone=true                          223ns ± 1%   237ns ± 0%   +6.28%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=false                          454ns ± 1%   445ns ± 0%   -2.01%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=2/two-level=true/bloom=true/with-tombstone=true                           261ns ± 0%   257ns ± 1%   -1.63%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=false                        249ns ± 1%   246ns ± 0%   -1.31%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=false/with-tombstone=true                         216ns ± 0%   219ns ± 0%   +1.52%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=false                         448ns ± 1%   437ns ± 0%   -2.42%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=false/bloom=true/with-tombstone=true                          257ns ± 0%   252ns ± 1%   -1.96%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=false                         290ns ± 1%   282ns ± 0%   -2.76%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=false/with-tombstone=true                          227ns ± 1%   232ns ± 0%   +2.24%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=false                          453ns ± 1%   436ns ± 0%   -3.64%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGENotFound/skip=4/two-level=true/bloom=true/with-tombstone=true                           261ns ± 1%   253ns ± 1%   -3.13%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=false                           295ns ± 1%   299ns ± 0%   +1.30%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=false/with-tombstone=true                            255ns ± 1%   272ns ± 1%   +6.86%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=false                            534ns ± 0%   553ns ± 1%   +3.53%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=false/bloom=true/with-tombstone=true                             364ns ± 0%   361ns ± 0%   -0.83%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=false                            340ns ± 1%   332ns ± 0%   -2.38%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=false/with-tombstone=true                             270ns ± 0%   278ns ± 0%   +2.98%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=false                             572ns ± 0%   569ns ± 0%   -0.59%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=1/two-level=true/bloom=true/with-tombstone=true                              386ns ± 1%   379ns ± 2%     ~     (p=0.114 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=false                           313ns ± 2%   321ns ± 1%   +2.75%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=false/with-tombstone=true                            277ns ± 1%   289ns ± 0%   +4.23%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=false                            565ns ± 0%   580ns ± 0%   +2.70%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=false/bloom=true/with-tombstone=true                             388ns ± 0%   378ns ± 1%   -2.56%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=false                            358ns ± 0%   347ns ± 0%   -2.92%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=false/with-tombstone=true                             286ns ± 0%   293ns ± 0%   +2.61%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=false                             594ns ± 0%   586ns ± 0%   -1.35%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=2/two-level=true/bloom=true/with-tombstone=true                              401ns ± 1%   392ns ± 0%   -2.30%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=false                           350ns ± 0%   347ns ± 0%   -0.88%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=false/with-tombstone=true                            307ns ± 0%   315ns ± 1%   +2.44%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=false                            597ns ± 1%   600ns ± 0%     ~     (p=0.486 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=false/bloom=true/with-tombstone=true                             415ns ± 2%   407ns ± 0%   -1.74%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=false                            389ns ± 3%   378ns ± 0%   -2.73%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=false/with-tombstone=true                             312ns ± 1%   323ns ± 0%   +3.55%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=false                             607ns ± 0%   616ns ± 0%   +1.43%  (p=0.029 n=4+4)
IteratorSeqSeekPrefixGEFound/skip=4/two-level=true/bloom=true/with-tombstone=true                              426ns ± 0%   423ns ± 1%     ~     (p=0.086 n=4+4)
IteratorSeqSeekGEWithBounds/two-level=false                                                                    360ns ± 2%   360ns ± 0%     ~     (p=0.314 n=4+4)
IteratorSeqSeekGEWithBounds/two-level=true                                                                     365ns ± 0%   360ns ± 0%   -1.28%  (p=0.029 n=4+4)
IteratorSeekGENoop/withLimit=false                                                                            33.9ns ± 0%  26.7ns ± 1%  -21.43%  (p=0.029 n=4+4)
IteratorSeekGENoop/withLimit=true                                                                             56.3ns ± 3%  55.7ns ± 0%     ~     (p=0.343 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward/seekprefix                                           25.8ms ± 2%  25.5ms ± 4%     ~     (p=0.343 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/forward/next                                                 2.20ms ± 0%  2.15ms ± 0%   -2.23%  (p=0.029 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@10/backward                                                     2.65ms ± 1%  2.62ms ± 2%     ~     (p=0.486 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward/seekprefix                                           26.0ms ± 1%  25.3ms ± 0%   -2.54%  (p=0.029 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/forward/next                                                 1.91ms ± 1%  1.87ms ± 1%   -2.00%  (p=0.029 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@50/backward                                                     2.20ms ± 1%  2.24ms ± 0%   +1.89%  (p=0.029 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward/seekprefix                                           25.3ms ± 0%  25.4ms ± 0%   +0.66%  (p=0.029 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/forward/next                                                  953µs ± 0%   956µs ± 0%     ~     (p=0.057 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@75/backward                                                     1.22ms ± 0%  1.25ms ± 0%   +2.51%  (p=0.029 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward/seekprefix                                          25.8ms ± 2%  25.4ms ± 0%     ~     (p=0.057 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/forward/next                                                 176µs ± 0%   182µs ± 0%   +3.54%  (p=0.029 n=4+4)
Iterator_RangeKeyMasking/range-keys-suffixes=@100/backward                                                     313µs ± 2%   314µs ± 1%     ~     (p=1.000 n=4+4)
IteratorScan/keys=100,r-amp=1,key-types=points-only                                                           5.20µs ± 1%  4.92µs ± 0%   -5.56%  (p=0.029 n=4+4)
IteratorScan/keys=100,r-amp=1,key-types=points-and-ranges                                                     5.30µs ± 1%  5.09µs ± 1%   -4.03%  (p=0.029 n=4+4)
IteratorScan/keys=100,r-amp=3,key-types=points-only                                                           7.83µs ± 0%  7.84µs ± 1%     ~     (p=0.571 n=4+4)
IteratorScan/keys=100,r-amp=3,key-types=points-and-ranges                                                     8.02µs ± 0%  8.02µs ± 0%     ~     (p=1.000 n=4+4)
IteratorScan/keys=100,r-amp=7,key-types=points-only                                                           11.6µs ± 2%  12.2µs ± 2%   +4.88%  (p=0.029 n=4+4)
IteratorScan/keys=100,r-amp=7,key-types=points-and-ranges                                                     11.8µs ± 1%  12.3µs ± 2%   +3.74%  (p=0.029 n=4+4)
IteratorScan/keys=100,r-amp=10,key-types=points-only                                                          14.4µs ± 2%  14.4µs ± 3%     ~     (p=0.886 n=4+4)
IteratorScan/keys=100,r-amp=10,key-types=points-and-ranges                                                    14.8µs ± 2%  14.8µs ± 1%     ~     (p=0.886 n=4+4)
IteratorScan/keys=1000,r-amp=1,key-types=points-only                                                          45.4µs ± 1%  40.4µs ± 1%  -11.04%  (p=0.029 n=4+4)
IteratorScan/keys=1000,r-amp=1,key-types=points-and-ranges                                                    46.8µs ± 2%  41.4µs ± 1%  -11.66%  (p=0.029 n=4+4)
IteratorScan/keys=1000,r-amp=3,key-types=points-only                                                          66.5µs ± 1%  60.3µs ± 1%   -9.35%  (p=0.029 n=4+4)
IteratorScan/keys=1000,r-amp=3,key-types=points-and-ranges                                                    66.7µs ± 2%  61.8µs ± 1%   -7.32%  (p=0.029 n=4+4)
IteratorScan/keys=1000,r-amp=7,key-types=points-only                                                          86.0µs ± 1%  90.5µs ± 6%     ~     (p=0.343 n=4+4)
IteratorScan/keys=1000,r-amp=7,key-types=points-and-ranges                                                    89.0µs ± 1%  93.3µs ± 3%   +4.93%  (p=0.029 n=4+4)
IteratorScan/keys=1000,r-amp=10,key-types=points-only                                                         99.2µs ± 2%  96.3µs ± 0%   -2.95%  (p=0.029 n=4+4)
IteratorScan/keys=1000,r-amp=10,key-types=points-and-ranges                                                    100µs ± 0%   102µs ± 1%   +1.83%  (p=0.029 n=4+4)
IteratorScan/keys=10000,r-amp=1,key-types=points-only                                                          448µs ± 1%   397µs ± 1%  -11.29%  (p=0.029 n=4+4)
IteratorScan/keys=10000,r-amp=1,key-types=points-and-ranges                                                    467µs ± 1%   409µs ± 1%  -12.34%  (p=0.029 n=4+4)
IteratorScan/keys=10000,r-amp=3,key-types=points-only                                                          642µs ± 1%   593µs ± 0%   -7.62%  (p=0.029 n=4+4)
IteratorScan/keys=10000,r-amp=3,key-types=points-and-ranges                                                    649µs ± 0%   605µs ± 0%   -6.83%  (p=0.029 n=4+4)
IteratorScan/keys=10000,r-amp=7,key-types=points-only                                                          852µs ± 2%   849µs ± 5%     ~     (p=0.486 n=4+4)
IteratorScan/keys=10000,r-amp=7,key-types=points-and-ranges                                                    867µs ± 3%   887µs ± 0%     ~     (p=0.343 n=4+4)
IteratorScan/keys=10000,r-amp=10,key-types=points-only                                                         945µs ± 0%   898µs ± 1%   -5.05%  (p=0.029 n=4+4)
IteratorScan/keys=10000,r-amp=10,key-types=points-and-ranges                                                   958µs ± 1%   946µs ± 1%     ~     (p=0.057 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=false/key-types=points-only             73.6ns ± 1%  71.3ns ± 1%   -3.10%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=false/key-types=points-and-ranges       75.0ns ± 1%  72.8ns ± 0%   -2.88%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=true/key-types=points-only              74.6ns ± 1%  72.5ns ± 0%   -2.85%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=1/value-blocks=true/key-types=points-and-ranges        76.0ns ± 0%  73.7ns ± 0%   -3.06%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=false/key-types=points-only              104ns ± 1%   101ns ± 0%   -2.53%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=false/key-types=points-and-ranges        108ns ± 2%   103ns ± 1%   -4.73%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=true/key-types=points-only               107ns ± 2%   102ns ± 0%   -4.26%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=3/value-blocks=true/key-types=points-and-ranges         108ns ± 1%   104ns ± 0%   -3.39%  (p=0.029 n=4+4)
IteratorScanNextPrefix/keysPerLevel=10/versions=1/ramp=7/value-blocks=false/key-types=points-only              133ns ± 1%   131ns ± 0%     ~     (p=0.057 n=4+4)
…